# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/kyc": {
      "post": {
        "operationId": "KycController_submitKYC",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitKycDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "KYC"
        ]
      }
    },
    "/api/v1/kyc/me": {
      "get": {
        "operationId": "KycController_getMyKYC",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "KYC"
        ]
      }
    },
    "/api/v1/kyc/me/step/{stepName}": {
      "put": {
        "operationId": "KycController_updateKYCStep",
        "parameters": [
          {
            "name": "stepName",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "KYC"
        ]
      }
    },
    "/api/v1/kyc/case/{caseId}": {
      "get": {
        "operationId": "KycController_getKYCCase",
        "parameters": [
          {
            "name": "caseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "KYC"
        ]
      }
    },
...
    "/api/v1/orders": {
      "post": {
        "operationId": "OrdersController_createOrder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create order from accepted quote (Internal - called by quotes service)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/me": {
      "get": {
        "operationId": "OrdersController_getMyOrders",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my orders (Buyer/Seller)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_getOrderById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order by ID",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/status": {
      "put": {
        "operationId": "OrdersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update order status (Seller/Admin)",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/orders/{id}/cancel": {
      "put": {
        "operationId": "OrdersController_cancelOrder",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel order (Buyer/Seller/Admin)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/documents": {
      "get": {
        "operationId": "OrdersController_getOrderDocuments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order documents",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/documents/proforma-invoice": {
      "post": {
        "operationId": "OrdersController_uploadProformaInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload proforma invoice (Seller)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/documents/tax-invoice": {
      "post": {
        "operationId": "OrdersController_uploadTaxInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload tax invoice (Seller)",
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/payments/order/{orderId}": {
      "get": {
        "operationId": "PaymentsController_getPaymentByOrderId",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get payment for order",
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/api/v1/disputes": {
      "post": {
        "operationId": "DisputesController_raiseDispute",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RaiseDisputeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Raise dispute (Buyer/Seller)",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/disputes/me": {
      "get": {
        "operationId": "DisputesController_getMyDisputes",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my disputes",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/disputes/{id}": {
      "get": {
        "operationId": "DisputesController_getDisputeById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get dispute by ID",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/disputes/{id}/evidence": {
      "post": {
        "operationId": "DisputesController_uploadEvidence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload evidence",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/disputes/{id}/resolve": {
      "put": {
        "operationId": "DisputesController_resolveDispute",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDisputeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resolve dispute (Admin only)",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/disputes/{id}/assign": {
      "put": {
        "operationId": "DisputesController_assignDispute",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign dispute to admin (Admin)",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/disputes/order/{orderId}": {
      "get": {
        "operationId": "DisputesController_getDisputesByOrder",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get disputes by order",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/disputes/order/{orderId}/window-status": {
      "get": {
        "operationId": "DisputesController_checkDisputeWindow",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check if dispute window is open",
        "tags": [
          "Disputes"
        ]
      }
    },
...
    "/api/v1/admin/kyc/case/{kycCaseId}": {
      "get": {
        "operationId": "KycAdminController_getKycCaseById",
        "parameters": [
          {
            "name": "kycCaseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin KYC"
        ]
      }
    },
...
  },
  "info": {
    "title": "B2B Marketplace API",
    "description": "Backend API for B2B Marketplace Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SubmitKycDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {}
      },
...
      "RaiseDisputeDto": {
        "type": "object",
        "properties": {}
      },
      "ResolveDisputeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}